projectile added, deflectable projectile
Projectiles added, if deflectable the player can deflect it with the sword slash

hitbox can delete its parent if deleteoriginoncontact is true, mainly used for projectiles but could have other uses. as of now, it only delete its parent not the origin as the origin may change due to the deflectable change.

when deflecting a projectile that is deflectable the origin and isenemybox is changed to that of the entity deflecting the projectile, so a fireball owned by a zombie that gets deflected by the player will now have the origin of the player (aka the player now owns it) and the isenemybox will match that of the players hitbox's isenemybox boolean.