Deflectable projectiles, global particle mngr/spawner, health bar
Projectiles are deflectable if in range of player, by clicking/tapping on the projectile.

particle effects will be added to a dict in global along with a function to spawn one into the level at a given position.

basic particle effect shown on hit of any entity

simple player stats menu added to level with working health bar